fix: standardPronunciation 상수와 관련하여 새파일을 만든다
