+++ src/constants.ts
@@ -258,105 +258,3 @@ export const 종성_알파벳_발음 = {
   ㅇ: 'ng',
   '': '',
 } as const;
-
-export const 음가가_없는_자음 = 'ㅇ';
-
-export const 한글_자모 = ['기역', '니은', '리을', '미음', '비읍', '시옷', '이응'] as const;
-export const 특별한_한글_자모 = ['디귿', '지읒', '치읓', '키읔', '티읕', '피읖', '히읗'] as const;
-export const 특별한_한글_자모의_발음 = {
-  ㄷ: 'ㅅ',
-  ㅈ: 'ㅅ',
-  ㅊ: 'ㅅ',
-  ㅌ: 'ㅅ',
-  ㅎ: 'ㅅ',
-  ㅋ: 'ㄱ',
-  ㅍ: 'ㅂ',
-} as const;
-
-export const 음의_동화_받침 = {
-  ㄷ: 'ㅈ',
-  ㅌ: 'ㅊ',
-  ㄹㅌ: 'ㅊ',
-} as const;
-
-export const ㄴㄹ이_덧나는_모음 = ['ㅑ', 'ㅕ', 'ㅛ', 'ㅠ', 'ㅣ', 'ㅒ', 'ㅖ'] as const;
-export const ㄴㄹ이_덧나서_받침_ㄴ_변환 = ['ㄱ', 'ㄴ', 'ㄷ', 'ㅁ', 'ㅂ', 'ㅇ'] as const;
-export const ㄴㄹ이_덧나서_받침_ㄹ_변환 = ['ㄹ'] as const;
-
-// 19항
-export const 자음동화_받침_ㄴ_변환 = ['ㅁ', 'ㅇ', 'ㄱ', 'ㅂ'] as const;
-
-// 18항
-export const 비음화_받침_ㅇ_변환 = ['ㄱ', 'ㄲ', 'ㅋ', 'ㄱㅅ', 'ㄹㄱ'] as const;
-export const 비음화_받침_ㄴ_변환 = ['ㄷ', 'ㅅ', 'ㅆ', 'ㅈ', 'ㅊ', 'ㅌ', 'ㅎ'] as const;
-export const 비음화_받침_ㅁ_변환 = ['ㅂ', 'ㅍ', 'ㄹㅂ', 'ㄹㅍ', 'ㅂㅅ'] as const;
-
-// 12항
-export const 발음변환_받침_ㅎ = ['ㅎ', 'ㄴㅎ', 'ㄹㅎ'] as const;
-export const 발음변환_받침_ㅎ_발음 = {
-  ㄱ: 'ㅋ',
-  ㄷ: 'ㅌ',
-  ㅈ: 'ㅊ',
-  ㅅ: 'ㅆ',
-} as const;
-export const 발음변환_첫소리_ㅎ = ['ㄱ', 'ㄹㄱ', 'ㄷ', 'ㅂ', 'ㄹㅂ', 'ㅈ', 'ㄴㅈ'] as const;
-export const 발음변환_첫소리_ㅎ_발음 = {
-  ㄱ: 'ㅋ',
-  ㄹㄱ: 'ㅋ',
-  ㄷ: 'ㅌ',
-  ㅂ: 'ㅍ',
-  ㄹㅂ: 'ㅍ',
-  ㅈ: 'ㅊ',
-  ㄴㅈ: 'ㅊ',
-} as const;
-
-// 9항, 10항, 11항
-export const 받침_대표음_발음 = {
-  ㄲ: 'ㄱ',
-  ㅋ: 'ㄱ',
-  ㄱㅅ: 'ㄱ',
-  ㄹㄱ: 'ㄱ',
-  ㅅ: 'ㄷ',
-  ㅆ: 'ㄷ',
-  ㅈ: 'ㄷ',
-  ㅊ: 'ㄷ',
-  ㅌ: 'ㄷ',
-  ㅍ: 'ㅂ',
-  ㅂㅅ: 'ㅂ',
-  ㄹㅍ: 'ㅂ',
-  ㄴㅈ: 'ㄴ',
-  ㄹㅂ: 'ㄹ',
-  ㄹㅅ: 'ㄹ',
-  ㄹㅌ: 'ㄹ',
-  ㄹㅁ: 'ㅁ',
-} as const;
-
-// 경음화 23항
-export const 된소리_받침_23항 = [
-  'ㄱ',
-  'ㄲ',
-  'ㅋ',
-  'ㄱㅅ',
-  'ㄹㄱ',
-  'ㄷ',
-  'ㅅ',
-  'ㅆ',
-  'ㅈ',
-  'ㅊ',
-  'ㅌ',
-  'ㅂ',
-  'ㅍ',
-  'ㄹㅂ',
-  'ㄹㅍ',
-  'ㅂㅅ',
-] as const;
-
-export const 된소리 = {
-  ㄱ: 'ㄲ',
-  ㄷ: 'ㄸ',
-  ㅂ: 'ㅃ',
-  ㅅ: 'ㅆ',
-  ㅈ: 'ㅉ',
-} as const;
-
-export const 어간_받침 = ['ㄴ', 'ㄴㅈ', 'ㅁ', 'ㄹㅁ', 'ㄹㅂ', 'ㄹㅌ'] as const;
